fixed bug for accepting decimal type

+++ application/forms.py
@@ -1,6 +1,6 @@
 from flask_wtf import FlaskForm
 from models import Users
-from wtforms import Form, StringField, PasswordField, IntegerField, RadioField
+from wtforms import Form, StringField, PasswordField, RadioField, DecimalField
 from wtforms.validators import InputRequired, Length, EqualTo
 
 
@@ -16,7 +16,7 @@ class RegisterForm(FlaskForm):
 
 class UploadForm(FlaskForm):
 	product = StringField('product', validators=[InputRequired()], render_kw={"placeholder": "Enter product name"})
-	price = IntegerField('price',validators=[InputRequired()], render_kw={"placeholder": "Price(Must be numbers)"})
+	price = DecimalField('price',validators=[InputRequired()], render_kw={"placeholder": "Price(Must be numbers)"})
 	address = StringField('address',validators=[InputRequired()], render_kw={"placeholder": "Grocery store's address"})
 	price_type = RadioField(choices=[('Regular','Regular'),('Special','Special')],default='Regular')
